+++ fuc/api/pycov.py
@@ -321,8 +321,9 @@ class CovFrame:
         Parameters
         ----------
         fn : str or file-like object
-            TSV file (zipped or unzipped). By file-like object, we refer to
-            objects with a :meth:`read()` method, such as a file handle.
+            TSV file (compressed or uncompressed). By file-like object, we
+            refer to objects with a :meth:`read()` method, such as a file
+            handle.
         compression : bool, default: False
             If True, use GZIP decompression regardless of filename.
 
@@ -753,7 +754,7 @@ class CovFrame:
         Parameters
         ----------
         fn : str
-            TSV file (zipped or unzipped).
+            TSV file (compressed or uncompressed).
         compression : bool, default: False
             If True, use the GZIP compression.
         """