+++ fuc/cli/vcf_filter.py
@@ -45,7 +45,7 @@ def create_parser(subparsers):
     )
     parser.add_argument(
         'vcf',
-        help='VCF file (zipped or unzipped).'
+        help='VCF file (compressed or uncompressed).'
     )
     parser.add_argument(
         '--expr',