+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Changelog
 * Add new command :command:`vcf-index`.
 * Add new method :meth:`pyvcf.has_chr_prefix`.
 * Add new command :meth:`common.update_chr_prefix`.
+* Update :meth:`pyvcf.slice` method to automatically handle the 'chr' prefix.
 
 0.26.0 (2021-10-24)
 -------------------